+++ src/Schema/Directives/CountDirective.php
@@ -39,6 +39,18 @@ directive @count(
   Apply scopes to the underlying query.
   """
   scopes: [String!]
+
+  """
+  The field to count.
+  counts only a specific field
+  """
+  field: String
+
+  """
+  distinct count.
+  exclude duplicated rows
+  """
+  distinct: Boolean
 ) on FIELD_DEFINITION
 GRAPHQL;
     }
@@ -52,6 +64,15 @@ GRAPHQL;
                     ->namespaceModelClass($modelArg)::query();
 
                 $this->makeBuilderDecorator($resolveInfo)($query);
+
+                if ($this->directiveArgValue('distinct') ?? false) {
+                    $query->distinct();
+                }
+
+                $field = $this->directiveArgValue('field');
+                if ($field) {
+                    return $query->count($field);
+                }
 
                 return $query->count();
             });