+++ CHANGELOG.md
@@ -9,6 +9,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- Add `distinct` and `field` args to `@count` directive https://github.com/nuwave/lighthouse/pull/2138
+
 ## v5.49.0
 
 ### Added